separates hardcoded modal from app function and adds it to Home page, route
