+++ client/src/App.js
@@ -3,21 +3,21 @@ import { ApolloClient, InMemoryCache, createHttpLink, ApolloProvider} from '@apo
 import { setContext } from '@apollo/client/link/context';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 
-import { useState, useEffect } from 'react'
-import {Button} from "react-bootstrap"
+// import { useState, useEffect } from 'react'
+// import {Button} from "react-bootstrap"
 import './App.css'
 
-// import Home from './pages/Home';
+import Home from './pages/Home';
 import SignUpForm from './pages/SignUpForm';
 import Login from './pages/Login';
 // import Profile from './pages/Profile';
 import Header from './components/Header/index';
 import Footer from './components/Footer/index';
-import Form from './components/Form/Form';
-import FormModal from './components/FormModal/FormModal'
+// import Form from './components/Form/Form';
+// import FormModal from './components/FormModal/FormModal'
 
 
-import Card from './components/Card/Card'
+// import Card from './components/Card/Card'
 import { ProtectedRoute } from './components/ProtectedRoute';
 
 const httpLink = createHttpLink({
@@ -45,56 +45,56 @@ const client = new ApolloClient({
 
 
 function App() {
-  const [ showModal, setShowModal ] = useState(false)
-  const [ cardDetails, setCardDetails ] = useState({
-    holiday: '',
-    senderName: '',
-    recipientName: '',
-    customMessage: '',
-    message: 'Happy Holidays',
-    messageType: '',
-    imageUpload: ''
-  })
+//   const [ showModal, setShowModal ] = useState(false)
+//   const [ cardDetails, setCardDetails ] = useState({
+//     holiday: '',
+//     senderName: '',
+//     recipientName: '',
+//     customMessage: '',
+//     message: 'Happy Holidays',
+//     messageType: '',
+//     imageUpload: ''
+//   })
 
-  const handleUpload = (event) => {
-    console.log(event.target.files[0].name)
-    setCardDetails({
-      ...cardDetails, 
-      imageUpload: URL.createObjectURL(event.target.files[0])
-    })
-}
+//   const handleUpload = (event) => {
+//     console.log(event.target.files[0].name)
+//     setCardDetails({
+//       ...cardDetails, 
+//       imageUpload: URL.createObjectURL(event.target.files[0])
+//     })
+// }
 
-  const handleInputChange = (event) => {
-    setCardDetails({
-      ...cardDetails,
-      [event.target.name]: event.target.value
-    })
-  }
+//   const handleInputChange = (event) => {
+//     setCardDetails({
+//       ...cardDetails,
+//       [event.target.name]: event.target.value
+//     })
+//   }
 
-  const handleShow = () => {
-    setShowModal(true)
-  }
+//   const handleShow = () => {
+//     setShowModal(true)
+//   }
 
-  useEffect(() => {
-    // setShowModal(true)
-  }, [])
+//   useEffect(() => {
+//     // setShowModal(true)
+//   }, [])
 
-  const closeModal = () => {
-    setShowModal(false)
-  }
+//   const closeModal = () => {
+//     setShowModal(false)
+//   }
 
-  const handleSubmit = (event) => {
-    event.preventDefault();
+//   const handleSubmit = (event) => {
+//     event.preventDefault();
 
-    setShowModal(false)
-    // Submit form data and redirect to page with the greeting card component displayed
-    setCardDetails({
-      ...cardDetails,
-      [event.target.name]: event.target.value
-    })
+//     setShowModal(false)
+//     // Submit form data and redirect to page with the greeting card component displayed
+//     setCardDetails({
+//       ...cardDetails,
+//       [event.target.name]: event.target.value
+//     })
 
-    console.log(cardDetails)  
-  }
+//     console.log(cardDetails)  
+//   }
 
   return (
     <ApolloProvider client={client}>
@@ -103,10 +103,10 @@ function App() {
       <Header />
       <div className="App-header">
       <Routes>
-      {/* <Route 
+          <Route 
             path="/"
             element={<Home />}
-          />*/}
+          />
          <Route 
             path="/login"
             element={<Login />}
@@ -123,7 +123,7 @@ function App() {
             path="/Profile/:username"
             element={<Profile />}
           /> */}
-        <Route
+        {/* <Route
           path="/"
           element={ 
             <>
@@ -145,7 +145,7 @@ function App() {
               </FormModal>
             </>
           }
-        />
+        /> */}
        <Route
           path="/protected"
           element={
